requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0.1
schedule>=1.2.0
